# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Exemplo",
        "description": "Esta é uma API de exemplo para o teste técnico",
        "contact": {
            "email": "seuemail@exemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/users": {
            "get": {
                "summary": "Listar todos os usuários",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "Lista de usuários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "summary": "Obter usuário por ID",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Usuário",
                "description": "Modelo que representa um usuário",
                "properties": {
                    "id": {
                        "description": "ID do usuário",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email do usuário",
                        "type": "string"
                    },
                    "age": {
                        "description": "Idade do usuário",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    }
}